Add directory redirection handling to cowboy_static
Directory accesses with no trailing slash will by default be redirected
(301) to their canonical slashed version. New handler option 'directory_index'
can be set to 'false' to disable, or to another status code such as 302.

Directory accesses with a trailing slash can be redirected (301) to an
index file. This is off by default. New handler option 'directory_index' can
be set to 'true' or a status code such as 302, for the default "index.html",
or to a binary string such as <<"index.php">>, or to a pair {Code, Filename}.